Rename Reservation related components
